Migrate to gradle based build
This change introduces Gradle as the build system. The purpose is to make it
easier to build, with less manual steps.

Basically new contributors need to:

* install the JDK
* clone the repo
* use IntelliJ (CE is sufficient) and import the `build.gradle` as
  project from existing sources file, check "auto-import" and
  "auto create source roots", use the included wrapper as recommended
* nothing. That's it

The project leverages the Gradle wrapper which makes the individual
projects' build-tool independent from a (possibily) globally installed
gradle version.

All manual build commands should begin with `./gradlew` (windows users
should use the `.bat` file.

### Available tasks:

* `./gradlew buildPlugin` - builds a zip file in `.gradle/distributions`
* `./gradlew check` - runs the tests
* `./gradlew runIDE` - launches the currently configured IDE version
  with the plugin and dependencies installed

## Versioning

Currently, the version is derived from the git tag, and `plugin.xml` is
automatically through the build script. The current configuration
modifies the `since` and `until` properties to match the
build-environment.

Of course it's possible to change, but it's also possible to
automatically publish the plugin (https://github.com/JetBrains/gradle-intellij-plugin#publishing-dsl,
for reference: https://github.com/cedricziel/idea-php-typo3-plugin/blob/master/.travis.yml#L36:L53), so that's perfectly fine, but also changeable
if you need to publish it manually.

## Building for different IDEA versions

The configuration relies on gradle properties to determine the current
IDEA and PHP Plugin version.

To switch the current version, comment out the current config in
`gradle.properties` and un-comment the desired version. IDEA will pickup
the change  and reconfigure the project. In travis it's mostly the same,
but environment variable-driven.

## Running and debugging

Use IDEA to execute the `runIde` gradle task to run a IDE. Use the same
task as debug-run and set your breakpoints. Use the `reload classes`
task to recompile and hot-reload.

When reloading fails, you might want to delete the `out` folder and / or
the `.gradle/classes` directory and reload again.